// components/charting_library/charting_library.min.d.ts
export declare type LanguageCode =
  | 'ar'
  | 'zh'
  | 'cs'
  | 'da_DK'
  | 'nl_NL'
  | 'en'
  | 'et_EE'
  | 'fr'
  | 'de'
  | 'el'
  | 'he_IL'
  | 'hu_HU'
  | 'id_ID'
  | 'it'
  | 'ja'
  | 'ko'
  | 'fa'
  | 'pl'
  | 'pt'
  | 'ro'
  | 'ru'
  | 'sk_SK'
  | 'es'
  | 'sv'
  | 'th'
  | 'tr'
  | 'vi';
// ...
export interface SingleBrokerMetaInfo {
  configFlags: BrokerConfigFlags;
  customNotificationFields?: string[];
  durations?: OrderDurationMetaInfo[];
}
// ...
export interface IBrokerConnectionAdapterHost {
  factory: IBrokerConnectionAdapterFactory;
  connectionStatusUpdate(status: ConnectionStatus, message?: string): void;
  defaultFormatter(symbol: string): Promise<IFormatter>;
  numericFormatter(decimalPlaces: number): Promise<IFormatter>;
  defaultContextMenuActions(
    context: ITradeContext,
    params?: DefaultContextMenuActionsParams,
  ): Promise<ActionMetaInfo[]>;
  defaultDropdownMenuActions(
    options?: Partial<DefaultDropdownActionsParams>,
  ): ActionMetaInfo[];
  floatingTradingPanelVisibility(): IWatchedValue<boolean>;
  domVisibility(): IWatchedValue<boolean>;
  silentOrdersPlacement(): IWatchedValue<boolean>;
  patchConfig(config: Partial<BrokerConfigFlags>): void;
  setDurations(durations: OrderDurationMetaInfo[]): void;
  orderUpdate(order: Order, isHistoryUpdate?: boolean): void;
  orderPartialUpdate(id: string, orderChanges: Partial<Order>): void;
  positionUpdate(position: Position, isHistoryUpdate?: boolean): void;
  positionPartialUpdate(id: string, positionChanges: Partial<Position>): void;
  tradeUpdate(trade: Trade, isHistoryUpdate?: boolean): void;
  tradePartialUpdate(id: string, tradeChanges: Partial<Trade>): void;
  executionUpdate(execution: Execution, isHistoryUpdate?: boolean): void;
  fullUpdate(): void;
  realtimeUpdate(symbol: string, data: TradingQuotes): void;
  plUpdate(positionId: string, pl: number): void;
  tradePLUpdate(tradeId: string, pl: number): void;
  equityUpdate(equity: number): void;
  domeUpdate(symbol: string, equity: DOMData): void;
  showOrderDialog<T extends PreOrder>(
    order: T,
    handler: (
      order: T,
      customFieldsResult?: CustomInputFieldsValues,
    ) => Promise<void>,
    focus?: OrderTicketFocusControl,
    options?: OrderDialogOptions,
  ): Promise<void>;
  showCancelOrderDialog(orderId: string, handler: () => void): Promise<void>;
  showCancelMultipleOrdersDialog(
    symbol: string,
    side: Side | undefined,
    qty: number,
    handler: () => void,
  ): Promise<void>;
  showCancelBracketsDialog(orderId: string, handler: () => void): Promise<void>;
  showCancelMultipleBracketsDialog(
    orderId: string,
    handler: () => void,
  ): Promise<void>;
  showClosePositionDialog(
    positionId: string,
    handler: () => void,
  ): Promise<void>;
  showReversePositionDialog(
    position: Position,
    handler: () => void,
  ): Promise<void>;
  showPositionBracketsDialog(
    position: Position | Trade,
    brackets: Brackets,
    focus: OrderTicketFocusControl | null,
    handler: (brackets: Brackets) => void,
  ): Promise<void>;
  showNotification(
    title: string,
    text: string,
    notificationType?: NotificationType,
  ): void;
  setButtonDropdownActions(descriptions: ActionMetaInfo[]): void;
  activateBottomWidget(): Promise<void>;
  showTradingProperties(): void;
  suggestedQty(): SuggestedQuantity;
  symbolSnapshot(symbol: string): Promise<QuotesBase>;
}
export interface IBrokerWithoutRealtime extends IBrokerCommon {
  subscribeDOME?(symbol: string): void;
  unsubscribeDOME?(symbol: string): void;
  cancelOrder(orderId: string, silently: boolean): Promise<void>;
  cancelOrders(
    symbol: string,
    side: Side | undefined,
    ordersIds: string[],
    silently: boolean,
  ): Promise<void>;
  reversePosition?(positionId: string, silently?: boolean): Promise<void>;
  closePosition(positionId: string, silently: boolean): Promise<void>;
  closeTrade?(tradeId: string, silently: boolean): Promise<void>;
  /**
   * @deprecated Brokers should always send PL and equity updates
   */
  subscribePL?(positionId: string): void;
  subscribeEquity?(): void;
  /**
   * @deprecated
   */
  unsubscribePL?(positionId: string): void;
  unsubscribeEquity?(): void;
}
export interface IBrokerTerminal extends IBrokerWithoutRealtime {
  subscribeRealtime(symbol: string): void;
  unsubscribeRealtime(symbol: string): void;
}
export declare type ResolutionString = string;
// ...
export interface RestBrokerMetaInfo {
  url: string;
  access_token: string;
}
export interface AccessListItem {
  name: string;
  grayed?: boolean;
}
export interface AccessList {
  type: 'black' | 'white';
  tools: AccessListItem[];
}
export interface NumericFormattingParams {
  decimal_sign: string;
}
export declare type AvailableSaveloadVersions = '1.0' | '1.1';
export interface Overrides {
  [key: string]: string | number | boolean;
}
export interface WidgetBarParams {
  details?: boolean;
  watchlist?: boolean;
  news?: boolean;
  watchlist_settings?: {
    default_symbols: string[];
    readonly?: boolean;
  };
}
export interface RssNewsFeedInfo {
  url: string;
  name: string;
}
export declare type RssNewsFeedItem = RssNewsFeedInfo | RssNewsFeedInfo[];
export interface RssNewsFeedParams {
  default: RssNewsFeedItem;
  [symbolType: string]: RssNewsFeedItem;
}
export interface NewsProvider {
  is_news_generic?: boolean;
  get_news(symbol: string, callback: (items: NewsItem[]) => void): void;
}
// ...
export interface ChartingLibraryWidgetOptions {
  container_id: string;
  datafeed: IBasicDataFeed | (IBasicDataFeed & IDatafeedQuotesApi);
  interval: ResolutionString;
  symbol: string;
  auto_save_delay?: number;
  autosize?: boolean;
  debug?: boolean;
  disabled_features?: string[];
  drawings_access?: AccessList;
  enabled_features?: string[];
  fullscreen?: boolean;
  height?: number;
  library_path?: string;
  locale: LanguageCode;
  numeric_formatting?: NumericFormattingParams;
  saved_data?: object;
  studies_access?: AccessList;
  study_count_limit?: number;
  symbol_search_request_delay?: number;
  timeframe?: string;
  timezone?: 'exchange' | Timezone;
  toolbar_bg?: string;
  width?: number;
  charts_storage_url?: string;
  charts_storage_api_version?: AvailableSaveloadVersions;
  client_id?: string;
  user_id?: string;
  load_last_chart?: boolean;
  studies_overrides?: StudyOverrides;
  customFormatters?: CustomFormatters;
  overrides?: Overrides;
  snapshot_url?: string;
  indicators_file_name?: string;
  preset?: 'mobile';
  time_frames?: TimeFrameItem[];
  custom_css_url?: string;
  favorites?: Favorites;
  save_load_adapter?: IExternalSaveLoadAdapter;
  loading_screen?: LoadingScreenOptions;
  settings_adapter?: ISettingsAdapter;
  theme?: ThemeName;
}
export interface TradingTerminalWidgetOptions
  extends ChartingLibraryWidgetOptions {
  brokerConfig?: SingleBrokerMetaInfo;
  restConfig?: RestBrokerMetaInfo;
  widgetbar?: WidgetBarParams;
  rss_news_feed?: RssNewsFeedParams;
  news_provider?: NewsProvider;
  brokerFactory?(
    host: IBrokerConnectionAdapterHost,
  ): IBrokerWithoutRealtime | IBrokerTerminal;
}
export declare type LayoutType =
  | 's'
  | '2h'
  | '2-1'
  | '2v'
  | '3h'
  | '3v'
  | '3s'
  | '4'
  | '6'
  | '8';
export declare type SupportedLineTools =
  | 'text'
  | 'anchored_text'
  | 'note'
  | 'anchored_note'
  | 'double_curve'
  | 'arc'
  | 'icon'
  | 'arrow_up'
  | 'arrow_down'
  | 'arrow_left'
  | 'arrow_right'
  | 'price_label'
  | 'flag'
  | 'vertical_line'
  | 'horizontal_line'
  | 'horizontal_ray'
  | 'trend_line'
  | 'trend_angle'
  | 'arrow'
  | 'ray'
  | 'extended'
  | 'parallel_channel'
  | 'disjoint_angle'
  | 'flat_bottom'
  | 'pitchfork'
  | 'schiff_pitchfork_modified'
  | 'schiff_pitchfork'
  | 'balloon'
  | 'inside_pitchfork'
  | 'pitchfan'
  | 'gannbox'
  | 'gannbox_square'
  | 'gannbox_fixed'
  | 'gannbox_fan'
  | 'fib_retracement'
  | 'fib_trend_ext'
  | 'fib_speed_resist_fan'
  | 'fib_timezone'
  | 'fib_trend_time'
  | 'fib_circles'
  | 'fib_spiral'
  | 'fib_speed_resist_arcs'
  | 'fib_channel'
  | 'xabcd_pattern'
  | 'cypher_pattern'
  | 'abcd_pattern'
  | 'callout'
  | 'triangle_pattern'
  | '3divers_pattern'
  | 'head_and_shoulders'
  | 'fib_wedge'
  | 'elliott_impulse_wave'
  | 'elliott_triangle_wave'
  | 'elliott_triple_combo'
  | 'elliott_correction'
  | 'elliott_double_combo'
  | 'cyclic_lines'
  | 'time_cycles'
  | 'sine_line'
  | 'long_position'
  | 'short_position'
  | 'forecast'
  | 'date_range'
  | 'price_range'
  | 'date_and_price_range'
  | 'bars_pattern'
  | 'ghost_feed'
  | 'projection'
  | 'rectangle'
  | 'rotated_rectangle'
  | 'ellipse'
  | 'triangle'
  | 'polyline'
  | 'curve'
  | 'regression_trend'
  | 'cursor'
  | 'dot'
  | 'arrow_cursor'
  | 'eraser'
  | 'measure'
  | 'zoom'
  | 'brush';
// ...
export declare type EmptyCallback = () => void;
export interface SubscribeEventsMap {
  toggle_sidebar: (isHidden: boolean) => void;
  indicators_dialog: EmptyCallback;
  toggle_header: (isHidden: boolean) => void;
  edit_object_dialog: (params: EditObjectDialogEventParams) => void;
  chart_load_requested: (savedData: object) => void;
  chart_loaded: EmptyCallback;
  mouse_down: (params: MouseEventParams) => void;
  mouse_up: (params: MouseEventParams) => void;
  drawing: (params: StudyOrDrawingAddedToChartEventParams) => void;
  study: (params: StudyOrDrawingAddedToChartEventParams) => void;
  undo: EmptyCallback;
  redo: EmptyCallback;
  reset_scales: EmptyCallback;
  compare_add: EmptyCallback;
  add_compare: EmptyCallback;
  'load_study template': EmptyCallback;
  onTick: (tick: Bar) => void;
  onAutoSaveNeeded: EmptyCallback;
  onScreenshotReady: (url: string) => void;
  onMarkClick: (markId: Mark['id']) => void;
  onTimescaleMarkClick: (markId: TimescaleMark['id']) => void;
  onSelectedLineToolChanged: EmptyCallback;
  layout_about_to_be_changed: (newLayoutType: LayoutType) => void;
  layout_changed: EmptyCallback;
  activeChartChanged: (chartIndex: number) => void;
}
export interface SaveChartToServerOptions {
  chartName?: string;
  defaultChartName?: string;
}
// ...
export interface GrayedObject {
  type: 'drawing' | 'study';
  name: string;
}
// ...
export interface IChartingLibraryWidget {
  onChartReady(callback: EmptyCallback): void;
  onGrayedObjectClicked(callback: (obj: GrayedObject) => void): void;
  onShortcut(shortCut: string, callback: EmptyCallback): void;
  subscribe<EventName extends keyof SubscribeEventsMap>(
    event: EventName,
    callback: SubscribeEventsMap[EventName],
  ): void;
  unsubscribe<EventName extends keyof SubscribeEventsMap>(
    event: EventName,
    callback: SubscribeEventsMap[EventName],
  ): void;
  chart(index?: number): IChartWidgetApi;
  setLanguage(lang: LanguageCode): void;
  setSymbol(
    symbol: string,
    interval: ResolutionString,
    callback: EmptyCallback,
  ): void;
  remove(): void;
  closePopupsAndDialogs(): void;
  selectLineTool(linetool: SupportedLineTools): void;
  selectedLineTool(): SupportedLineTools;
  save(callback: (state: object) => void): void;
  load(state: object): void;
  getSavedCharts(callback: (chartRecords: SaveLoadChartRecord[]) => void): void;
  loadChartFromServer(chartRecord: SaveLoadChartRecord): void;
  saveChartToServer(
    onComplete?: EmptyCallback,
    onFail?: EmptyCallback,
    saveAsSnapshot?: false,
    options?: SaveChartToServerOptions,
  ): void;
  removeChartFromServer(
    chartId: string,
    onCompleteCallback: EmptyCallback,
  ): void;
  onContextMenu(
    callback: (unixTime: number, price: number) => ContextMenuItem[],
  ): void;
  createButton(options?: CreateButtonOptions): JQuery;
  showNoticeDialog(params: DialogParams<() => void>): void;
  showConfirmDialog(params: DialogParams<(confirmed: boolean) => void>): void;
  showLoadChartDialog(): void;
  showSaveAsChartDialog(): void;
  symbolInterval(): SymbolIntervalResult;
  mainSeriesPriceFormatter(): IFormatter;
  getIntervals(): string[];
  getStudiesList(): string[];
  addCustomCSSFile(url: string): void;
  applyOverrides<TOverrides extends StudyOverrides>(
    overrides: TOverrides,
  ): void;
  applyStudiesOverrides(overrides: object): void;
  watchList(): WatchListApi;
  activeChart(): IChartWidgetApi;
  chartsCount(): number;
  layout(): LayoutType;
  setLayout(layout: LayoutType): void;
  changeTheme(themeName: ThemeName): void;
  takeScreenshot(): void;
  lockAllDrawingTools(): IWatchedValue<boolean>;
}
export interface ChartingLibraryWidgetConstructor {
  new (
    options: ChartingLibraryWidgetOptions | TradingTerminalWidgetOptions,
  ): IChartingLibraryWidget;
}
export declare function version(): string;
export declare function onready(callback: () => void): void;
export declare const widget: ChartingLibraryWidgetConstructor;